It's Alive! Added Paste Model and Views
This commit contains the Paste model, which uses the postgresql testing
database to store and retrieve pastes. I also added a couple new view
elements to accomodate the new functionality.

There were a couple changes to the routes, as I realized that some of
the functionality similarities between the regular paste and secure
paste retrieval could be re-written to use the same code.

Curretly, when a new Paste is created, there is no validation to make
sure that the key it generates doesn't collide with another existing
key, and no code to account for the errors that might happen if it did.
That is something that I plan to add in the next commit, but I wanted to
get all these modified files pushed so that I could start fine tuning
each aspect separately.

File Modifications:
	modified:   app/controllers/Application.java
	new file:   app/models/Paste.java
	new file:   app/views/paste.scala.html
	new file:   app/views/showPaste.scala.html
	modified:   app/views/textbox.scala.html
	modified:   conf/application.conf
	modified:   conf/routes