Updated Paste Model
I've added just the smallest bit of verification that checks if the id
generated for a new paste already exists, and if it does, it tries to
generate a new one.

I have also added a couple comments in the code, but nothing life
changing. This is a pretty minor commit.

File Modifications:
	modified:   app/models/Paste.java