Bug Fix for Pastes With No Key
There was a small bug that I let slip in where a paste that had no key
would cause an error if someone attempted to give a key for it. While
fixing it, I decided to extend the implementation to just ignore all
keys given to pastes that do not require one.

File Modifications:
	modified:   app/models/Paste.java