Route Tweeks
I changed some of the route definitions to be more narrowly defined so
that they match the ids and keys I'm generating. In an ideal world, I'd
add some 404 and catch-all routing for everything that doesn't fit into
those boxes, but right now I'm not too concerned about it.

File Modifications:
	modified:   conf/routes