Visual Rework and Cleanup
So now that things work more or less how I want them to, I cleaned up a
few of the css and html things that weren't as nice as I would have
liked.

This properly renames the navi into a header, since it's really not
navigational, it's just the document top. I've also added a div for the
main content portion as well.

File Modifications:
	new file:   app/views/header.scala.html
	modified:   app/views/main.scala.html
	deleted:    app/views/navi.scala.html
	modified:   app/views/showPaste.scala.html
	modified:   app/views/textbox.scala.html
	modified:   public/stylesheets/playpaste.css